🔨 파일구조 정리
